Separate elements into own files + misc
