+++ src/elements/mod.rs
@@ -1,0 +1,34 @@
+use std::{path::PathBuf, time::SystemTime};
+
+use agdb::{DbElement, DbError, DbId, DbKeyValue, DbUserValue, DbValue, QueryId, UserValue};
+
+use crate::nodetype::{NodePhysicality, TypeName};
+
+pub (crate) mod node;
+pub (crate) mod node_path;
+pub (crate) mod edge;
+pub (crate) mod attribute;
+
+
+#[derive(Debug, Clone)]
+pub struct SysTime(SystemTime);
+
+impl From<SysTime> for DbValue {
+    fn from(time: SysTime) -> Self {
+        time.0.duration_since(SystemTime::UNIX_EPOCH).unwrap().as_secs().into()
+    }
+}
+
+impl TryFrom<DbValue> for SysTime {
+    type Error = DbError;
+
+    fn try_from(value: DbValue) -> Result<Self, Self::Error> {
+        Ok(SysTime(SystemTime::UNIX_EPOCH + std::time::Duration::from_secs(value.to_u64().unwrap())))
+    }
+}
+
+
+
+
+
+
