+++ tests/node.rs
@@ -222,7 +222,7 @@
 
 //     // NOTE: The below assertions are commented out because the amount of 
 //     // nodes created at startup is not set in stone. The most recent breakage 
-//     // was because of adding a nodecategories node. No point in breaking a test 
+//     // was because of adding a nodetypes node. No point in breaking a test 
 //     // every time that happens. 
 
 //     // let edges = edges.elements.iter().filter(|x| x.id.0 < 0).collect::<Vec<_>>();
